+++ products/dated-irs/src/modules/ProductIRSModule.sol
@@ -33,7 +33,7 @@ contract ProductIRSModule is IProductIRSModule {
     /**
      * @inheritdoc IProductIRSModule
      */
-    function initiateTakerOrder(uint128 accountId, uint128 marketId, uint256 maturityTimestamp, int256 baseAmount)
+    function initiateTakerOrder(uint128 accountId, uint128 marketId, uint32 maturityTimestamp, int256 baseAmount)
         external
         override
         returns (int256 executedBaseAmount, int256 executedQuoteAmount)
@@ -58,7 +58,7 @@ contract ProductIRSModule is IProductIRSModule {
      * @inheritdoc IProductIRSModule
      */
 
-    function settle(uint128 accountId, uint128 marketId, uint256 maturityTimestamp) external override {
+    function settle(uint128 accountId, uint128 marketId, uint32 maturityTimestamp) external override {
         Portfolio.Data storage portfolio = Portfolio.load(accountId);
         int256 settlementCashflowInQuote = portfolio.settle(marketId, maturityTimestamp);
